requirements.txt
console-menu==0.5.1
python-binance==0.7.1
python-telegram-bot==12.0.0b1
color==0.1.1
matplotlib==3.0.3
